+++ server/app.js
@@ -9,8 +9,7 @@ const bodyParser = require('body-parser');
 const errorHandler = require(path.resolve(__dirname, './utils/errorHandling'));
 const { stream } = require('./utils/logger');
 const connectDB = require('./database');
-const userRouter = require('./routes/v1.routes/userRouter');
-const secureRoute = require('./routes/v1.routes/secure-routes');
+const apiV1Router = require('./routes/v1.routes');
 require('./auth/auth');
 
 async function createApp(config) {
@@ -21,8 +20,10 @@ async function createApp(config) {
   app.use(morgan('combined', { stream }));
   app.use(bodyParser.urlencoded({ extended: false }));
 
-  app.use('/', userRouter);
-  app.use('/user', passport.authenticate('jwt', { session: false }), secureRoute);
+  app.use('/apiv1', apiV1Router);
+
+  //use later for secure routes
+  //app.use('/user', passport.authenticate('jwt', { session: false }), secureRoute);
 
   app.use('*', (req, res) => {
     res.status(404).send('Not Found');
