--- server/routes/v1.routes/userRouter.js
@@ -1,45 +1,0 @@
-const express = require('express');
-const passport = require('passport');
-const jwt = require('jsonwebtoken');
-const router = express.Router();
-
-router.post(
-  '/signup',
-  passport.authenticate('signup', { session: false }),
-  async (req, res, next) => {
-    res.json({
-      message: 'Signup successful',
-      user: req.user,
-    });
-  }
-);
-
-router.post('/login', async (req, res, next) => {
-  passport.authenticate('login', async (err, user, info) => {
-    try {
-      if (err || !user) {
-        // console.log(err)
-        const error = {
-          log: 'incorrect login',
-          status: 400,
-          message: { err: 'An error occurred' },
-        };
-
-        return next(error);
-      }
-
-      req.login(user, { session: false }, async (error) => {
-        if (error) return next(error);
-
-        const body = { _id: user._id, username: user.username };
-        const token = jwt.sign({ user: body }, 'TOP_SECRET');
-
-        return res.json({ token });
-      });
-    } catch (error) {
-      return next(error);
-    }
-  })(req, res, next);
-});
-
-module.exports = router;
